+++ agileflow-backend/src/main/java/com/agileflow/agileflow_backend/sprint/service/impl/SprintServiceImpl.java
@@ -12,19 +12,27 @@ import com.agileflow.agileflow_backend.sprint.service.SprintService;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-
+import com.agileflow.agileflow_backend.notification.service.NotificationService;
+import com.agileflow.agileflow_backend.issue.repository.IssueRepository;
+import com.agileflow.agileflow_backend.common.enums.NotificationType;
 @Service
 public class SprintServiceImpl implements SprintService {
 
     private final SprintRepository sprintRepository;
     private final ProjectRepository projectRepository;
+    private final NotificationService notificationService;
+    private final IssueRepository issueRepository;
 
     public SprintServiceImpl(
             SprintRepository sprintRepository,
-            ProjectRepository projectRepository) {
+            ProjectRepository projectRepository,
+            NotificationService notificationService,
+            IssueRepository issueRepository) {
 
         this.sprintRepository = sprintRepository;
         this.projectRepository = projectRepository;
+        this.notificationService = notificationService;
+        this.issueRepository = issueRepository;
     }
 
     @Override
@@ -54,6 +62,20 @@ public class SprintServiceImpl implements SprintService {
 
         sprint = sprintRepository.save(
                 sprint);
+
+        notificationService.create(
+
+                project.getOwner(),
+
+                "Sprint Created",
+
+                "Sprint " + sprint.getName() + " created",
+
+                NotificationType.SPRINT_CREATED,
+
+                "/sprints/" + sprint.getId()
+
+        );
 
         return map(
                 sprint);
@@ -112,7 +134,19 @@ public class SprintServiceImpl implements SprintService {
 
         sprint = sprintRepository.save(
                 sprint);
+        notificationService.create(
 
+                sprint.getProject().getOwner(),
+
+                "Sprint Updated",
+
+                "Sprint " + sprint.getName() + " updated",
+
+                NotificationType.SPRINT_UPDATED,
+
+                "/sprints/" + sprint.getId()
+
+        );
         return map(
                 sprint);
 
@@ -128,6 +162,24 @@ public class SprintServiceImpl implements SprintService {
                         .orElseThrow(() ->
                                 new ResourceNotFoundException(
                                         "Sprint not found"));
+
+        if (issueRepository.existsBySprintId(id)) {
+            throw new IllegalArgumentException("Sprint contains issues. Delete issues first.");
+        }
+
+        notificationService.create(
+
+                sprint.getProject().getOwner(),
+
+                "Sprint Deleted",
+
+                "Sprint " + sprint.getName() + " deleted",
+
+                NotificationType.SPRINT_DELETED,
+
+                "/sprints"
+
+        );
 
         sprintRepository.delete(
                 sprint);
